refactor: Phase 7 finalizer and debug cleanup (#168)
Split finalize-brief.ts into focused modules:
- normalize-editorial-input.ts
- prepare-debug-context.ts
- hydrate-debug-context.ts
- render-outputs.ts

Rename debug types to slot roles (primary/fallback):
- DebugGeminiDiagnostics -> DebugPrimaryDiagnostics
- DebugGroqDiagnostics -> DebugFallbackDiagnostics
- Add slot role fields to DebugAiTrace

Clean up editorial resolution types:
- Remove weekStandoutParseStatus from base payload
- Add EditorialCandidatePayloadWithCompat for boundaries
- Mark provider-specific types as deprecated

Update affected tests and documentation.
