+++ src/adapters/n8n/format-messages.adapter.test.ts
@@ -11,6 +11,7 @@ import {
   resolveSpurSuggestion,
   shouldReplaceAiText,
 } from '../../domain/editorial/resolution/resolve-editorial.js';
+import { parseEditorialResponseWithCompat } from '../../domain/editorial/resolution/parse.js';
 import {
   run,
 } from './format-messages.adapter.js';
@@ -654,7 +655,7 @@ describe('parseGroqResponse — backward compatibility (deprecated)', () => {
 
   it('reports present status and raw value when weekStandout is in the response', () => {
     const raw = JSON.stringify({ editorial: 'Good.', composition: [], weekStandout: 'Wednesday is the standout day.' });
-    const result = parseGroqResponse(raw);
+    const result = parseEditorialResponseWithCompat(raw);
     expect(result.weekStandoutParseStatus).toBe('present');
     expect(result.weekStandoutRawValue).toBe('Wednesday is the standout day.');
     expect(result.weekInsight).toBe('Wednesday is the standout day.');
@@ -662,7 +663,7 @@ describe('parseGroqResponse — backward compatibility (deprecated)', () => {
 
   it('reports absent status when weekStandout is missing from a valid JSON response', () => {
     const raw = JSON.stringify({ editorial: 'Good.', composition: [] });
-    const result = parseGroqResponse(raw);
+    const result = parseEditorialResponseWithCompat(raw);
     expect(result.weekStandoutParseStatus).toBe('absent');
     expect(result.weekStandoutRawValue).toBeNull();
     expect(result.weekInsight).toBe('');
@@ -670,13 +671,13 @@ describe('parseGroqResponse — backward compatibility (deprecated)', () => {
 
   it('reports absent status when weekStandout is not a string', () => {
     const raw = JSON.stringify({ editorial: 'Good.', composition: [], weekStandout: 42 });
-    const result = parseGroqResponse(raw);
+    const result = parseEditorialResponseWithCompat(raw);
     expect(result.weekStandoutParseStatus).toBe('absent');
     expect(result.weekStandoutRawValue).toBeNull();
   });
 
   it('reports parse-failure status when the response is not valid JSON', () => {
-    const result = parseGroqResponse('Not JSON at all.');
+    const result = parseEditorialResponseWithCompat('Not JSON at all.');
     expect(result.weekStandoutParseStatus).toBe('parse-failure');
     expect(result.weekStandoutRawValue).toBeNull();
     expect(result.weekInsight).toBe('');
