+++ src/presenters/email/debug-email.test.ts
@@ -84,6 +84,9 @@ describe('formatDebugEmail', () => {
       ai: {
         primaryProvider: 'gemini',
         selectedProvider: 'gemini',
+        rawFallbackResponse: '{"editorial":"Good night."}',
+        rawPrimaryResponse: '{"editorial":"Good night.","composition":["Low northern horizon"]}',
+        rawPrimaryPayload: '{"candidates":[{"content":{"parts":[{"text":"{\\"editorial\\":\\"Good night.\\"}","thoughtSignature":"sig"}]}}],"usageMetadata":{"thoughtsTokenCount":261}}',
         rawGroqResponse: '{"editorial":"Good night."}',
         rawGeminiResponse: '{"editorial":"Good night.","composition":["Low northern horizon"]}',
         rawGeminiPayload: '{"candidates":[{"content":{"parts":[{"text":"{\\"editorial\\":\\"Good night.\\"}","thoughtSignature":"sig"}]}}],"usageMetadata":{"thoughtsTokenCount":261}}',
@@ -226,6 +229,9 @@ describe('formatDebugEmail — new debug sections', () => {
         selectedProvider: 'groq',
         primaryRejectionReason: 'response truncated (MAX_TOKENS); editorial validation failed: editorial must contain two sentences',
         secondaryRejectionReason: null,
+        rawFallbackResponse: '{"editorial":"Conditions hold through the astro window.","composition":["Face north with a low tree line"]}',
+        rawPrimaryResponse: '{"editorial":"The moon sets before the midnight astro window begins at 00:',
+        rawPrimaryPayload: '{"body":{"data":{"candidates":[{"content":{"parts":[{"text":"truncated"}]}}]}}}',
         rawGroqResponse: '{"editorial":"Conditions hold through the astro window.","composition":["Face north with a low tree line"]}',
         rawGeminiResponse: '{"editorial":"The moon sets before the midnight astro window begins at 00:',
         rawGeminiPayload: '{"body":{"data":{"candidates":[{"content":{"parts":[{"text":"truncated"}]}}]}}}',
